More thoughts on move generation/representation

+++ move_t.cpp
@@ -1,4 +1,4 @@
 #include "move_t.h"
 
-move_t::move_t(size_t from, size_t to)
+move_t::move_t(int from, int to)
 : m_from(from), m_to(to) {}
